ps2/n64: sra is a 32-bit shift — the 64-bit one clipped every glyph in the game
draw-string computed correct pens and never drew a character. The glyph-emit code
was reached ~1300 times a frame and bailed every single time, one instruction
before the transform, on this:

	qmfc2 $t5, $vf1     # t5 = the vector: lane x in bits 31:0, lane y in 63:32
	bltz  $t5, done     # bltz reads bit 63 -> tests lane Y
	sra   $t5, $t5, 31  # bring lane X's sign into the register's sign
	bltz  $t5, next     # -> tests lane X

That is the EE's standard "test two vector lanes" idiom, and `sra 31` exists in it
for exactly one purpose: to move bit 31 into the sign so the second bltz can read
it. Our sra shifted the whole 64-bit register and took bit 63 as the sign, so the
second test dragged lane y's mantissa down into the sign bit and answered about the
wrong lane. Both lanes held comfortably positive floats (limit - pen = +883, +214);
we clipped anyway. Every string in the game, on every screen, drew nothing.

MIPS defines sra as rt[31:0] shifted with bit 31 replicated, sign-extended to 64 —
the high word is never read. Reading it is a dsra. Fixed in both 64-bit cores.

The mistake was load-bearing in three places and defended by all of them:

  - r4300 had it first and r5900 was ported from it, so the R5900's diff test —
    which checks the two cores against each other — passed. Its `interesting()`
    generator even draws values whose high half is not the sign extension of the
    low, specifically to expose this. Both cores were wrong together, exactly the
    caveat that test's own header warns about.
  - r4300's TestArithmeticShiftReadsTheWholeRegister asserted the wrong rule
    outright, deriving its expectation from the implementation's reasoning rather
    than from hardware.
  - Both cores carried the same confident copy-pasted comment arguing for it.

Nothing external ever checked it. What settles it: the MIPS III/64 pseudocode
(temp <- (GPR[rt]_31)^s || GPR[rt]_31..s); tools/cpu/mips, the R3000A core that IS
checked against an external vector set, which shifts int32(rt); and shipped game
code that only runs because silicon ignores the high half. The manual calls rt
UNPREDICTABLE when it is not a sign-extended word, which is how an implementer
talks themselves into either rule.

TestSRAIgnoresHighWord now pins it against the architecture rather than a sibling,
using the exact lanes draw-string fails on. r4300's test is corrected to the real
rule. Pilotwings 64's harnesses and the PS2 suite are unchanged and green.

RESULT: the memory-card prompt renders its text, centred and legible — "NO MEMORY
CARD (PS2) INSERTED IN MEMORY CARD SLOT 1 ... CONTINUE? / OK".

New instrument: -logvf ADDR:VF[,VF...] logs VU0's vector registers at a PC. -logpc
shows GPRs, but a routine computing in COP2 macro mode keeps its whole state in the
VFs, where every other instrument is blind; -bp + -vu0regs stops at the first hit
and cannot watch a value evolve across a loop. It named this bug in one run: the
pen sat still and the limits stayed positive while the code kept branching away.

+++ tools/cpu/r4300/exec.go
@@ -224,18 +224,20 @@ func (c *CPU) special(w, rs, rt uint32) {
 	case 0x02: // srl
 		c.set(rd, sext32(uint32(c.R[rt])>>shamt))
 	case 0x03: // sra
-		// The arithmetic right shifts read the *whole* 64-bit register, not its
-		// low half, and only then sign-extend the 32-bit result. A core that
-		// shifts the low half alone gives a different answer whenever the high
-		// half is not already the sign extension of the low one — which is
-		// exactly what happens after a 64-bit operation.
-		c.set(rd, sext32(uint32(int64(c.R[rt])>>shamt)))
+		// sra is a 32-BIT operation on a 64-bit register: MIPS defines the result as
+		// rt[31:0] shifted right with bit 31 replicated into the vacated bits, then
+		// sign-extended to 64. The upper half is never read — reading it, and taking
+		// bit 63 as the sign, is a dsra. (The manual calls rt UNPREDICTABLE when it
+		// is not a sign-extended word, but the silicon simply ignores the high half,
+		// which is what code relying on it expects. The 32-bit R3000A core in
+		// tools/cpu/mips, which is checked against an external vector set, agrees.)
+		c.set(rd, sext32(uint32(int32(uint32(c.R[rt]))>>shamt)))
 	case 0x04: // sllv
 		c.set(rd, sext32(uint32(c.R[rt])<<(c.R[rs]&31)))
 	case 0x06: // srlv
 		c.set(rd, sext32(uint32(c.R[rt])>>(c.R[rs]&31)))
 	case 0x07: // srav
-		c.set(rd, sext32(uint32(int64(c.R[rt])>>(c.R[rs]&31))))
+		c.set(rd, sext32(uint32(int32(uint32(c.R[rt]))>>(c.R[rs]&31))))
 
 	case 0x08: // jr
 		c.doBranch(true, c.R[rs])
